Practica algoritmo Codewars

+++ algoritmo17.js
@@ -1,5 +1,7 @@
 // Multiplicar
 
+// Este código no se ejecuta correctamente. Trate de averiguar por qué.
+
 const multiply = (a, b)=> {
   
     return a * b
@@ -8,4 +10,7 @@ const multiply = (a, b)=> {
 console.log(multiply(1,1))
 console.log(multiply(2,1))
 console.log(multiply(2,2))
-console.logl(multiply(3,5))
+console.log(multiply(3,5))
+
+// Interpretación
+// Se multiplica los 2 valores a * b
